feat(collector): move decodePem to shared util
